+++ src/metaobject-repository.ts
@@ -2,21 +2,23 @@ import { GraphQLClient } from "node_modules/@shopify/shopify-app-remix/dist/ts/s
 import { AdminOperations } from "@shopify/admin-api-client";
 import { FieldBuilder, QueryBuilder } from "raku-ql";
 import { camel } from "snake-camel";
-import { Collection, Company, Customer, GenericFile, Job, MediaImage, Metaobject, MetaobjectBulkDeletePayload, MetaobjectCreatePayload, MetaobjectDeletePayload, MetaobjectsCreatePayload, MetaobjectUpdatePayload, MetaobjectUpsertPayload, Page, PageInfo, Product, ProductVariant, TaxonomyValue, Video } from "~/types/admin.types";
+import { Collection, Company, Customer, GenericFile, Job, MediaImage, Metaobject, MetaobjectBulkDeletePayload, MetaobjectCreatePayload, MetaobjectDeletePayload, MetaobjectsCreatePayload, MetaobjectStatus, MetaobjectUpdatePayload, MetaobjectUpsertPayload, Page, PageInfo, Product, ProductVariant, TaxonomyValue, Video } from "~/types/admin.types";
 import { DefinitionSchema, DefinitionSchemaEntry, FieldDefinition, FromDefinitionWithSystemData, ValidPopulatePaths } from "./types/definitions";
 import { CreateInput, FindOptions, OnPopulateFunc, PopulateOptions, SortKey, UpdateInput, UpsertInput } from "./types/metaobject-repository";
 import { UserErrorsException } from "./exception/user-errors-exception";
 import { deserialize, serializeFields } from "./transformer";
 
-/**
- * Make a specific key in an object nullable
- */
-type WithNullableKey<T, K extends keyof T> = {
-  [P in keyof T]:
-    P extends K
-      ? T[P] | null
-      : T[P]
-};
+type EmptyObject<T extends { system: object }> =
+  // 2) Remove the original `system`
+  Omit<T, "system"> & {
+    // 3) Re-add `system` with all keys except `capabilities` made nullable
+    system: {
+      [K in keyof T["system"]]:
+        K extends "type" | "capabilities"
+          ? T["system"][K]
+          : T["system"][K] | null;
+    };
+  };
 
 /**
  * Object repository
@@ -40,12 +42,32 @@ export class MetaobjectRepository<
   /**
    * Generate a new empty object that contains all the fields of the definition, set to empty value
    */
-  getEmptyObject(): WithNullableKey<FromDefinitionWithSystemData<D, T>, 'system'> {
+  getEmptyObject(opts?: { defaultPublishableStatus: MetaobjectStatus }): EmptyObject<FromDefinitionWithSystemData<D, T>> {
+    const definition = this.getDefinitionSchemaEntry(this.type);
+
     let data: any = {
-      system: null
+      system: {
+        id: null,
+        handle: null,
+        type: this.type,
+        displayName: '',
+        createdAt: null,
+        updatedAt: null,
+        thumbnail: null
+      }
     };
 
-    const definition = this.getDefinitionSchemaEntry(this.type);
+    Object.keys(definition.capabilities || {}).forEach(capability => {
+      if (capability === 'publishable') {
+        data.system.capabilities.publishable = {
+          status: opts?.defaultPublishableStatus || 'draft'
+        }
+      } else if (capability === 'onlineStore') {
+        data.system.capabilities.onlineStore = {
+          templateSuffix: ''
+        }
+      }
+    })
 
     definition.fields.forEach((field) => {
       data[camel(field.key)] = field.type.startsWith('list.') ? [] : null;